Makes mobile an optional parameter in /notify/positive endpoint
If a mobile number is not provided then an SMS will not be sent. If a mobile number is provided and is valid but an error occurs while sending an SMS then the code will still be returned to the caller with an indication that the SMS was not sent.